tests/wpt/web-platform-tests/css/CSS2/normal-flow/unresolvable-max-height.html

<!DOCTYPE html>
<title>Unresolvable percentage min-height</title>
<link rel="author" title="Morten Stenshorne" href="mstensho@chromium.org">
<link rel="help" href="https://www.w3.org/TR/CSS22/visudet.html#min-max-heights" title="10.7 Minimum and maximum heights: 'min-height' and 'max-height'">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<div id="container" style="width:100px; height:30000px; max-height:100%;" data-expected-height="30000">
  <div style="height:12345px;"></div>
</div>
<script>
  checkLayout("#container");
</script>
